added all separated styled components from form to one styles file
