--- src/components/form/formStyles/textWrapper.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components'
-
-const TextWrapper = styled.div`
-  @media screen and (min-width: 768px) {
-    display: flex;
-    margin: 0;
-  }
-`
-
-export default TextWrapper
