+++ src/redux/auth/authActions.js
@@ -8,10 +8,6 @@ const loginRequest = createAction('auth/loginRequest')
 const loginSuccess = createAction('auth/loginSuccess')
 const loginError = createAction('auth/loginError')
 
-const googleRequest = createAction('auth/googleRequest')
-const googleSuccess = createAction('auth/googleSuccess')
-const googleError = createAction('auth/googleError')
-
 const logoutRequest = createAction('auth/logoutRequest')
 const logoutSuccess = createAction('auth/logoutSuccess')
 const logoutError = createAction('auth/logoutError')
@@ -19,8 +15,6 @@ const logoutError = createAction('auth/logoutError')
 const getCurrentUserRequest = createAction('auth/getCurrentUserRequest')
 const getCurrentUserSuccess = createAction('auth/getCurrentUserSuccess')
 const getCurrentUserError = createAction('auth/getCurrentUserError')
-
-const isModalShow = createAction('auth/isModalShow')
 
 export default {
   registerRequest,
@@ -35,8 +29,4 @@ export default {
   getCurrentUserRequest,
   getCurrentUserSuccess,
   getCurrentUserError,
-  googleRequest,
-  googleSuccess,
-  googleError,
-  isModalShow,
 }
